refactor(activemodel,activerecord): split TS-only naming tests, converge enum/removeColumns/schema-creation (RFC 0115, RFC 0051)
